# Restructure builder/ → core.topology + core.state_ops + graph_assembly/
Dissolves the `nodes/base.py` ↔ `builder/__init__.py` circular import by
splitting `fabricpc/builder/` along its two real responsibilities
(declarative topology primitives vs. the heavy graph assembly function),
promoting each to its proper layer in the dependency DAG. Also removes a dead
file, and consolidates `GraphState` helpers in `core/`.

## Moves

### `fabricpc/builder/edge.py` → `fabricpc/core/topology.py`
- `class SlotRef`
- `class Edge`

### `fabricpc/builder/namespace.py` → `fabricpc/core/topology.py`
- `_namespace_stack` (module-level `threading.local`)
- `def _get_current_namespace`
- `class GraphNamespace`

### `fabricpc/builder/graph_construction.py` → `fabricpc/graph_assembly/graph_construction.py`
- `class TaskMap`
- `def _build_slots`
- `def _topological_sort` — now the single canonical implementation; the
  duplicate `GraphStructure._topological_sort` in `core/types.py` was
  removed in a prior step
- `def graph`

Internal import updated:
```
from fabricpc.builder.edge import Edge, SlotRef
→ from fabricpc.core.topology import Edge, SlotRef
```

### `fabricpc/utils/helpers.py` → `fabricpc/core/state_ops.py`
- `def update_node_in_state`
- `def set_latents_to_clamps`

(`def layernorm` stays in `utils/helpers.py` — generic JAX math, used by
`nodes/transformer_v2.py`.)

## New files

- **`fabricpc/core/topology.py`** — declarative topology primitives. No
  `fabricpc` deps; safe to import from any layer including `nodes/`.
- **`fabricpc/core/state_ops.py`** — `GraphState` manipulation helpers.
- **`fabricpc/graph_assembly/__init__.py`** — re-exports `graph`, `TaskMap`.

## Deleted

- `fabricpc/builder/__init__.py`
- `fabricpc/builder/edge.py`
- `fabricpc/builder/namespace.py`
- `fabricpc/builder/graph_construction.py`
- `fabricpc/builder/` (package)
- `fabricpc/graph_initialization/graph_net.py` — dead; file body was
  `"""remove this file"""`

## Updated

### `fabricpc/nodes/base.py`
Promoted two function-local lazy imports to the file header (the only
remaining "real circular" carve-out in the codebase):

```
__init__: from fabricpc.builder.namespace import _get_current_namespace
slot()  : from fabricpc.builder.edge import SlotRef
→ from fabricpc.core.topology import SlotRef, _get_current_namespace
  (at module top)
```

### `fabricpc/__init__.py`
- Replaced submodule re-export `builder` with `graph_assembly`.
- Reordered submodule loads so `nodes` precedes `graph_assembly`
  (`nodes/transformer_v2.py` imports back from `graph_assembly` at module
  top; loading `nodes` first keeps `nodes.base` fully initialized before
  `graph_assembly/__init__.py` runs `graph_construction`).
- Updated example in module docstring:

```
from fabricpc.builder import Edge, TaskMap, graph
→ from fabricpc.core.topology import Edge
  from fabricpc.graph_assembly import TaskMap, graph
```

### `fabricpc/utils/__init__.py`
Dropped re-exports of the moved state helpers; `layernorm` stays.

## Callsite migrations (~30 files in `examples/`, `tests/`, `scripts/`, `nodes/`)

```
from fabricpc.builder import Edge, TaskMap, graph
→ from fabricpc.core.topology import Edge
  from fabricpc.graph_assembly import TaskMap, graph
```

```
from fabricpc.builder import Edge, TaskMap, graph, GraphNamespace
→ from fabricpc.core.topology import Edge, GraphNamespace
  from fabricpc.graph_assembly import TaskMap, graph
```
(`examples/mupc_demo.py` only)

```
from fabricpc.utils.helpers import set_latents_to_clamps
→ from fabricpc.core.state_ops import set_latents_to_clamps
```

```
from fabricpc.utils.helpers import update_node_in_state
→ from fabricpc.core.state_ops import update_node_in_state
```

## Layering after restructure (deps strictly upward)

```
core/types, core/topology, core/state_ops             (foundation)
  → core/{initializers, activations, energy, scaling, positional, mupc}
  → core/{inference, learning}
  → nodes/                                            (depends only on core/)
  → graph_assembly/                                   (depends on nodes + core)
  → graph_initialization/, training/, tuning/, experiments/
```

## Verification

- `import fabricpc.{nodes.base, graph_assembly}` in either order: clean.
- `pytest tests/ -x`: 127 passed.
- Demos run clean: `python examples/mnists_demo.py` and `python examples/transformer_v2_demo.py`.
