+++ README.md
@@ -70,7 +70,8 @@ A model consists of structure and parameters.
 
 ```python
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.inference import InferenceSGD
