+++ docs/user_guides/02_quickstart.md
@@ -21,7 +21,8 @@ set_jax_flags_before_importing_jax(jax_platforms="cuda")
 
 import jax
 from fabricpc.nodes import Linear, IdentityNode
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SigmoidActivation, SoftmaxActivation
 from fabricpc.core.energy import CrossEntropyEnergy
