+++ examples/custom_node.py
@@ -24,7 +24,8 @@ from fabricpc.utils.data.dataloader import MnistLoader
 
 from fabricpc.nodes import Linear
 from fabricpc.nodes.base import NodeBase, SlotSpec
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     IdentityActivation,
