+++ examples/mnist_aim_tracking.py
@@ -25,7 +25,8 @@ import jax.numpy as jnp
 from fabricpc.utils.data.dataloader import MnistLoader
 
 from fabricpc.nodes import Linear, IdentityNode
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params, FeedforwardStateInit
 from fabricpc.core.activations import (
     IdentityActivation,
