+++ examples/mnist_demo.py
@@ -26,7 +26,8 @@ set_jax_flags_before_importing_jax(
 
 import jax
 from fabricpc.nodes import Linear, IdentityNode
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SigmoidActivation, SoftmaxActivation
 from fabricpc.core.energy import CrossEntropyEnergy
