+++ examples/mnist_lateral_connections.py
@@ -33,7 +33,8 @@ import jax
 import argparse
 
 from fabricpc.nodes import Linear, IdentityNode
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     SigmoidActivation,
