+++ examples/mupc_demo.py
@@ -50,7 +50,8 @@ import optax
 
 from fabricpc.nodes import Linear, IdentityNode, LinearResidual
 from fabricpc.nodes.skip_connection import SkipConnection
-from fabricpc.builder import Edge, TaskMap, graph, GraphNamespace
+from fabricpc.core.topology import Edge, GraphNamespace
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import TanhActivation, SoftmaxActivation
 from fabricpc.core.energy import CrossEntropyEnergy
