+++ examples/resnet18_cifar10_demo.py
@@ -55,7 +55,8 @@ from custom_node import Conv2DNode
 from fabricpc.nodes import Linear, IdentityNode, SkipConnection
 from fabricpc.nodes.base import NodeBase, SlotSpec
 from fabricpc.core.types import NodeParams
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.core import InferenceSGDNormClip
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
