+++ examples/scaling/mlp_scaling.py
@@ -55,7 +55,8 @@ from dataclasses import dataclass, asdict
 from typing import List, Dict, Tuple, Optional
 
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params, FeedforwardStateInit
 from fabricpc.core.activations import IdentityActivation, SigmoidActivation
 from fabricpc.core.inference import InferenceSGD
