+++ examples/storkey_hopfield_demo.py
@@ -59,7 +59,8 @@ import jax
 import optax
 
 from fabricpc.nodes import Linear, IdentityNode, StorkeyHopfield
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core import TanhActivation
 from fabricpc.core.activations import SoftmaxActivation
