+++ examples/storkey_hopfield_diagnostic.py
@@ -34,7 +34,8 @@ import jax.numpy as jnp
 import optax
 
 from fabricpc.nodes import Linear, IdentityNode, StorkeyHopfield
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SoftmaxActivation, TanhActivation
 from fabricpc.core.energy import CrossEntropyEnergy, GaussianEnergy
@@ -43,7 +44,7 @@ from fabricpc.core.initializers import XavierInitializer
 from fabricpc.training import train_pcn, evaluate_pcn
 from fabricpc.training.train import train_step
 from fabricpc.graph_initialization.state_initializer import initialize_graph_state
-from fabricpc.utils.helpers import update_node_in_state
+from fabricpc.core.state_ops import update_node_in_state
 from fabricpc.utils.data.dataloader import (
     FashionMnistLoader,
     FewShotLoader,
