+++ fabricpc/graph_assembly/__init__.py
@@ -1,0 +1,5 @@
+"""Graph assembly: turn declared nodes and edges into a GraphStructure."""
+
+from fabricpc.graph_assembly.graph_construction import graph, TaskMap
+
+__all__ = ["graph", "TaskMap"]
