--- fabricpc/graph_initialization/graph_net.py
@@ -1,3 +1,0 @@
-"""
-remove this file
-"""
