+++ tests/test_auto_node_grad.py
@@ -15,7 +15,8 @@ from fabricpc.nodes import (
     Linear,
     LinearExplicitGrad,
 )
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import (
     IdentityActivation,
