+++ tests/test_energy.py
@@ -18,7 +18,8 @@ from fabricpc.core.energy import (
     KLDivergenceEnergy,
 )
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.inference import InferenceSGD
