+++ tests/test_ndim_shapes.py
@@ -11,7 +11,8 @@ import jax
 import jax.numpy as jnp
 
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.core.inference import InferenceSGD
