+++ tests/test_optimizers.py
@@ -14,7 +14,8 @@ from fabricpc.training.optimizers import (
 )
 from fabricpc.training import train_step
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import SigmoidActivation
 from fabricpc.core.inference import InferenceSGD
