+++ tests/test_state_initializer.py
@@ -12,7 +12,8 @@ import jax.numpy as jnp
 
 from fabricpc.nodes import Linear
 from fabricpc.nodes.transformer import TransformerBlock
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.inference import InferenceSGD, run_inference
 from fabricpc.core.activations import (
