+++ tests/test_storkey_hopfield.py
@@ -5,7 +5,8 @@ import jax
 import jax.numpy as jnp
 
 from fabricpc.nodes import Linear, IdentityNode, StorkeyHopfield
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.graph_initialization.state_initializer import initialize_graph_state
 from fabricpc.core.learning import compute_local_weight_gradients
