+++ tests/test_train_backprop.py
@@ -9,7 +9,8 @@ import jax.numpy as jnp
 import optax
 
 from fabricpc.nodes import Linear
-from fabricpc.builder import Edge, TaskMap, graph
+from fabricpc.core.topology import Edge
+from fabricpc.graph_assembly import TaskMap, graph
 from fabricpc.graph_initialization import initialize_params
 from fabricpc.core.activations import ReLUActivation, SoftmaxActivation
 from fabricpc.core.inference import InferenceSGD
